autorefresh standings!

+++ app/controllers/player_pool_controller.rb
@@ -3,7 +3,7 @@ require 'uri'
 
 class PlayerPoolController < ApplicationController
   def index
-    @users = User.all.sort_by{|u| u.total_points}.reverse
+    @users = User.includes(:players => [:boxscores, :team]).sort_by{|u| u.total_points}.reverse
   end
 
   def refresh
@@ -13,12 +13,14 @@ class PlayerPoolController < ApplicationController
     raise "ERROR: HTTP RESPONSE STATUS NOT 200: #{http_response.body}" if http_response.code != '200'
     doc = Nokogiri::HTML(http_response.body)
     doc.search('#ysp-leaguescoreboard table.ysptblclbg3').each do |game|
-      link = game.search('table').last.search('.yspscores').last.search('a').first['href']
-      url = "http://rivals.yahoo.com/#{link}"
+      link = game.search('table').last.search('.yspscores').last.search('a').first
+      next unless link.inner_html == "Box Score"
+      url = "http://rivals.yahoo.com/#{link['href']}"
       game = Game.find_by_url url
       game.destroy if game && !game.is_final
       Game.create(:url => url) if game.nil? || game.destroyed?
     end
-    redirect_to root_url
+    @users = User.includes(:players => [:boxscores, :team]).sort_by{|u| u.total_points}.reverse
+    render :partial => 'standings'
   end
 end
